Add files via upload
Found these code in internet, and try to run it on my Command Prompt, but it couldn't open. My Windows CMD's Turtle was NOT RESPONDING, 
Is it due low RAM ? I only have 8GB RAM on my laptop. Maybe it's too poor to run Python codes? Or did I miss a line to makes the code runs. Advise will be a HUGE help here. Thank you.